feat: validate bindings
