+++ pkg/apis/v1alpha1/wait.go
@@ -4,30 +4,6 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
-// For specifies the condition to wait for.
-type For struct {
-	// Deletion specifies parameters for waiting on a resource's deletion.
-	// +optional
-	Deletion *Deletion `json:"deletion,omitempty"`
-
-	// Condition specifies the condition to wait for.
-	// +optional
-	Condition *Condition `json:"condition,omitempty"`
-}
-
-// Deletion represents parameters for waiting on a resource's deletion.
-type Deletion struct{}
-
-// Condition represents parameters for waiting on a specific condition of a resource.
-type Condition struct {
-	// Name defines the specific condition to wait for, e.g., "Available", "Ready".
-	Name string `json:"name"`
-
-	// Value defines the specific condition status to wait for, e.g., "True", "False".
-	// +optional
-	Value *string `json:"value,omitempty"`
-}
-
 // Wait specifies how to perform wait operations on resources.
 type Wait struct {
 	// Timeout for the operation. Specifies how long to wait for the condition to be met before timing out.
