+++ pkg/apis/v1alpha1/wait_deletion.go
@@ -1,0 +1,4 @@
+package v1alpha1
+
+// Deletion represents parameters for waiting on a resource's deletion.
+type Deletion struct{}
