+++ pkg/validation/test/error.go
@@ -9,6 +9,7 @@ func ValidateError(path *field.Path, obj *v1alpha1.Error) field.ErrorList {
 	var errs field.ErrorList
 	if obj != nil {
 		errs = append(errs, ValidateFileRefOrCheck(path, obj.FileRefOrCheck)...)
+		errs = append(errs, ValidateBindings(path.Child("bindings"), obj.Bindings...)...)
 	}
 	return errs
 }
