+++ pkg/validation/test/script.go
@@ -10,8 +10,9 @@ func ValidateScript(path *field.Path, obj *v1alpha1.Script) field.ErrorList {
 	if obj != nil {
 		if obj.Content == "" {
 			errs = append(errs, field.Invalid(path.Child("content"), obj, "content must be specified"))
-			errs = append(errs, ValidateCheck(path.Child("check"), obj.Check)...)
 		}
+		errs = append(errs, ValidateCheck(path.Child("check"), obj.Check)...)
+		errs = append(errs, ValidateBindings(path.Child("bindings"), obj.Bindings...)...)
 	}
 	return errs
 }
